[feature] 优化对焦流程 [why] none [how] none [influence] 对焦速度，清晰度。 [jira]

+++ app/src/main/java/com/cvte/camera2demo/util/OpenCVUtils.java
@@ -25,7 +25,7 @@ public class OpenCVUtils {
      */
     public static double calculateOneBitmapClarityWithNoGray(Bitmap srcBitmap) {
         int kernel_size = 3;
-        int ddepth = CvType.CV_16U;
+        int ddepth = CvType.CV_8U;
         Mat mat = new Mat();
         Utils.bitmapToMat(srcBitmap, mat);
         mat = cutImgROI(mat);
